wrong type of int for count

+++ 10-binary_tree_depth.c
@@ -15,7 +15,7 @@ size_t binary_tree_depth(const binary_tree_t *tree)
 		if (binary_tree_is_root(tree) == 1)
 			return (0);
 
-		int left_depth, right_depth;
+		size_t left_depth, right_depth;
 
 		left_depth = binary_tree_depth(tree->left);
 		right_depth = binary_tree_depth(tree->right);
